fixed hidden bug at reading XML config values in UWP build of framework

+++ 3FD/configuration.cpp
@@ -321,7 +321,7 @@ namespace _3fd
 				return defaultValue;
 
 			auto innerText = nodeConfig->InnerText->Data();
-			long parsedVal = wcstod(innerText, nullptr);
+			float parsedVal = wcstod(innerText, nullptr);
 			return parsedVal > 0.0 ? parsedVal : defaultValue;
 		}
